% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windfetch_class.R
\name{as_sf,WindFetch-method}
\alias{as_sf,WindFetch-method}
\title{Coerce a WindFetch object to an sf Object}
\usage{
\S4method{as_sf}{WindFetch}(x)
}
\arguments{
\item{x}{a WindFetch object created using the \link{windfetch} function.}
}
\description{
Coerce a WindFetch object to an sf Object.
}
