% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windfetch_class.R
\name{as_sf}
\alias{as_sf}
\title{Coerce a WindFetch object to an sf Object}
\usage{
as_sf(x)
}
\arguments{
\item{x}{a WindFetch object created using the \link{windfetch} function.}
}
\description{
Coerce a WindFetch object to an sf Object.
}
