% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windfetch_class.R
\name{asTibble}
\alias{asTibble}
\title{Coerce a WindFetch Object to a Tibble}
\usage{
asTibble(x)
}
\arguments{
\item{x}{a WindFetch object created using the \link{windfetch} function.}
}
\description{
Coerce a WindFetch object to a tibble.
}
