% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windfetch_class.R
\name{crs_transform}
\alias{crs_transform}
\title{Transform the CRS of a WindFetch object}
\usage{
crs_transform(x, y)
}
\arguments{
\item{x}{a WindFetch object created using the \link{windfetch} function.}

\item{y}{character string denoting the EPSG code in the format "epsg:xxxx".
This parameter is passed into the \link[sf]{st_transform} function to
do the transformation.}
}
\description{
Transform the CRS of a WindFetch object.
}
