% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windfetch_class.R
\name{plot,WindFetch,missing-method}
\alias{plot,WindFetch,missing-method}
\title{Plot a WindFetch Object}
\usage{
\S4method{plot}{WindFetch,missing}(x, y, ...)
}
\arguments{
\item{x}{a \code{\link{WindFetch}} object as returned by
\code{\link{windfetch}}.}

\item{y}{missing.}

\item{...}{further arguments passed to
\code{\link{plot}}.}
}
\description{
Plot method for \code{\link{WindFetch}} objects.
}
