% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windfetch_class.R
\name{summary,WindFetch-method}
\alias{summary,WindFetch-method}
\alias{show,WindFetch-method}
\title{Summarise a WindFetch Object}
\usage{
\S4method{summary}{WindFetch}(object)

\S4method{show}{WindFetch}(object)
}
\arguments{
\item{object}{a \code{WindFetch} object that has been returned by the
\code{\link{windfetch}} function.}
}
\value{
The \code{summary} function returns a \code{\link[dplyr]{tibble}}.
}
\description{
The \code{summary} function calculates the average wind fetch for the separate
northerly, easterly, southerly and westerly quadrants. For example, the mean
fetch for the northerly component averages over the fetch vectors with
directions between 315 (inclusive) and 45 (exclusive) degrees, i.e. the fetch
vectors within the interval [315, 45).
}
