% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windfetch_class.R
\docType{class}
\name{WindFetch}
\alias{WindFetch}
\title{WindFetch Class}
\description{
Class to hold WindFetch objects.
}
\section{Slots}{

\describe{
\item{\code{names}}{character vector containing the names for each location.}

\item{\code{max_dist}}{numeric vector of length 1 containing the maximum distance (km)
of the fetch vectors.}
}}

\note{
WindFetch objects should only be created using the \code{\link{windfetch}}
constructor function.
}
